requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
sqlalchemy-utils
passlib[bcrypt]
python-jose
requests